Some cleanup, exposed and fixed a bug w/r/t ''this''.
''this'' is not always an object, since the function it is calling
might be a non-object's built-in method.  (All variable types are at
least partly objects.  I should research how other interpreters
do this, because my way is sort of unclean.)  So type checking needs
to be done before de-referencing return value of get_this().