Re-write array code.
This is largely untested.  Arrays for floats, ints, and strings do not
store the whole struct qvar_t.  This is more RAM friendly, but it comes
at the cost of code cleanliness and uniformity, because calling function
has to provide the return-value struct var_t as an argument to get a
data element from the array, rather than having the function just return
the element.  For now I can live with that.