Replace 'children' field of object handle with a hash table.
This doesn't speed things up, but it doesn't slow things down,
either.  If I add a per-frame hash table, then maybe the removal
of a variable's 'name' field (reducing it to 24 bytes instead of
32) might speed things up, idunno.

At the moment, my test of non-internal sqrt is about 20% the speed
of the same algorithm using Python.
I'm not embarrassed exactly, but I am disappointed.