Go back to manual argument parsing.
The everywhere standard version of getopt does not permit '-' options
after the first non-option argument (but why?), but when the shell runs
a script as an executable, it apparently passes the non-optional arg
(name of the script) first.  Unless I'm mistaken, but that's the behavior
I was seeing. So manually parse options.