Fixed break statement issue w/r/t stack unwinding...
...but only to find a problem with breaking from the "else" of a
"for-else" block.  I traded a bigger problem for a lesser problem, so
it's a net win, I quess.  I have a solution for the "else", but it's a
a rat's nest of B instructions to ensure that PUSH/POP_BLOCK is balanced
in every case scenario.  I'd rather wait and see if I can think of
something better before I commit to it.