Start of re-shifting responsibility of loading a file.
If input is TTY, I want to execute every time a full expression
has been received (plus enter, since TTY's often line buffer,
but I'm ok with that).  assemble_expression() in assemble.c is a
good wrapping point for that, but the code wrapping it is all wrong,
and the file-by-file API needs to rethink that.  Meanwhile, have
assemble.c call prescan(), I'll figure out the rest later.

This commit also deletes some dead code in evilcandy.h, references
to functions in a non-existent compile.c