remove lexer struct, replace it with something more re-entrant.
This is part of a migration of the parser-state API out from lex.c.
The idea will be to get rid of prescan() and have its return value,
the token buffer, grow only grow as needed by as_lex/as_unlex in
assembler.c.  This is so eventually EvilCandy can operate in interactive
mode; currently it parses all the way to EOF before it starts executing.