Make string be pass-by-value instead of pass-by-reference.
This is just the implementation change, the documentation and comments
elsewhere pertaining to this will be updated in another commit.

Having strings be pass-by-reference is faster and uses less RAM,
but that's the only advantage against a sea of disadvantages.
Trying out how Python and JavaScript treat strings ("text objects"
in Python), it's clear that no one else uses pass-by-reference strings.