VM's callbacks now return a value.
This slows the VM down a tiny bit and the the callbacks just return zero
for now, but it allows me to add proper runtime exception handling. Current
error handling is too blunt: EvilCandy prints an error message and quits.
Work-arounds are cringe: E.G. if the Io module fails its open() method,
it returns a string with an error message instead of a file object.