Fix string-not-set bug found by occasional crashes from demo/cat.py.
When string_get_from_file (or whatever it's called) reads an empty line,
its buffer does not get set, and somewhere downstream something accesses
it while NULL and causes a seg fault.

Just always stuff a nullchar into the buffer when creating a string
object, so there's always a buffer to de-reference.