Resolve DRY violation between do_call_func and vm_reenter, and call internal and external functions both from the same place.
The biggest trade-off here is recursion.  Previously, I had been
staying within the same call of execute_loop() when a user-defined
function has been called, so that recursion only occurs when a new
script is being loaded or a built-in C function further calls a user
callback.  This made a speghettified mess of things, and now I've
finally given up on it.  execute_loop() recurses for _every_ nested
function call, but look how much simpler and cleaner the
call_function() interface and vm_reenter/do_call_function are.