Overhaul the API for built-in function callbacks.
I hoped this would greatly simplify things, but it's a bit of a meh
when all is said and done.  It's a big lift, so there are bound to be
some little bugs inserted here and there than need weeding out.
Good thing no one is payimg me to do this:)