Overhauled the VM/assembler's way of dealing with variable assignment.
This is part of a bigger overhaul, to
1. make the memory management more manageable,
2. get rid of the hated qop_mov() function
3. simplify the way we access the global name resolution.

This is currently ten steps backward for #3.  The symbol table
seeking and setting functions are more complicated now, and they
have a lot of redundant duplicate searches in case we try to replace
something.  This will be fixed in a later commit.

But it is a tiny step forward for #1 and a huge leap forward for #2.