BIG LIFT: Embed struct var_t in each typedef's metadata struct.
Now there's only one malloc()/free() when creating a variable, only
one dereferece to get to the variable's metadata.  This update also gets
rid of StrptrType in the process.

This was a huge change so even though my demos all run without a hiccup,
it undoubtedly entered some bugs in places and caused a lot of dead code
that needs to be removed.  Imagine how disappointed I am, then, to find
that it's making no measure improvement on performance.