Have all frames share the same global (for now, while still single-threaded) stack.
Before this commit, each frame had a static frame size, so no function
can grow the stack bigger and many functions didn't use it all.  So it
was both limiting _and_ inefficient.  The assembler could calculate how
much stack usage a given function will use and add that to the build-
function opcode arg, but a global stack is better anyway, because it
improves locality of reference.