Remove LIST_APPEND instruction.
Now that we have a bigger stack, put the list items on the stack
until we get to the end of the expression, then give DEFLIST a
'number of items' argument, and have the instruction build the
list all at once.  This is more optimal because the array_new
can now take a size argument and not require repetitive use of
array_append at this stage.  (Technically this is still happening,
but it's now limited to array.c's internals, which will be fixed
soon.)