Check if end-of-input before calling assemble().
If assemble() returns due to EOF, and it's not an unexpected EOF,
program terminated correctly.  But for some reason if the EOF was
unexpected and assemble() returned an error due to that, the next
call to assemble() would just hang, requiring ^C to quit the program.
I could spelunk assemble.c looking for the cause, but it's easier
just to not call assemble() again if stdin has reached EOF.