Fix one bug in closure parsing and log some others in to-do.txt
In some cases, the token state a->oc had been updated when checking for
a symbol.  This is fine in most cases, because a 'name' arg was used
which copied the earlier a->oc.  However, when the symbol is found to
be a closure, the parser needed to evaluate the closure at the parent
level, and since the token position had changed, bugs happened.
as_savetok now saves both the token and the token position.

Other bugs persist w/r/t closures, these were added to to-do.txt.