Replace 'load' keyword with 'import' built-in function, which behaves differently.
import takes two arguments: file name (or module name if library module),
and mode.  If mode is 'r', then compile the file and return it as a function
handle.  If mode is 'x', then compile and execute the file and return its
result.  This reduces the need for adding global variables in modules, because
they can just return to the caller whatever information needs to be passed.