Moved parsing of string token from token stage to 'put in .rodata' stage.
This looks wasteful, but it will be meaningful when I add literals for
the bytes data type.  I don't want to have to figure out stuff like
backslash-zeros in the middle of a string during the tokenizing and
assembling phase.  Let the data-type handlers and *only* the data-type
handlers worry about that.