Do not make a new string var for string_cp; just produce a reference.
Strings are immutable now, have been for some time.  So there's no
longer a need to copy them from one variable to the next.  A sequence
like:
        a = 'foo';
        b = a;
        b += ' bar';
will not change a because the + operation will create a new string
and replace b's old string with the new one.

Additionally, I deleted a stray unused declaration that was causing
a compiler warning.