Delete .cp callback for typedefs.
This was the last remnant of the hated 'qop_mov' function, from when
everything was getting memcpy'd from one stack place to another.  Now,
everything is immutable (integers, strings...) or documented to be
by-reference (functions, dictionaries...).  So it makes no sense to
ever create a new instance in order to 'copy' one's self; instead just
produce another reference.