Be more standard-C compliant.
Lets be honest, Gnu **is** the standard now, but to pretend to be
portable, I will occasionally compile with '-std=c17 -pedantic-errors' to
see where I added code that I didn't realize was Gnu-only.  In this case,
'standard' C variadic macros are broken whereas Gnu's are not.  So I
added DBUG1 as a janky work-around DBUG messages whose strings take no
formatting arguments.