Get rid of SHIFT_DOWN solution of 'how do we know who "this" is?'
I don't *like* the solution, but maybe the real answer is to add
a class syntax after all.  The way we're doing it, we're creating
and destroying a variable quite frequently whenever we get a
function from an object, even if it isn't a per-se "method".
But there were hazards with the SHIFT_DOWN method, where a method
that's put into a separate variable (say, for convenience) no longer
has a reference to its owner.