Prevent accidental cyclic reference.
When a function is retrieved from a dictionary, the actual data is a
wrapping 'method', which points to the dictionary.  The high-level
user doesn't see this, they just think they got the function.  If the
function--now a method--is inserted again to the same dictionary,
regardless what the new key is, take it out of the method and insert
the parentless function.  Other objects may own the method, so that
'this' is always for the first dictionary that contained the function.