Cleaner way to get properties.
Property getters/setters are wrapped in a PropertyType object, which is
then put in the same dictionary as the type's built-in methods.  When
retrieved, var_getattr() calls the getter directly and returns that.

There is so far no var_setattr() version of any of this.  So far all
objects' built-in properties are read-only.