Make interactive mode more interactive by printing results automatically.
No longer need to use print() to get output while in interactive mode.
If in interactive mode, and the statement was an evaluatable expression
(ie. 'x;' instead of 'x = y;'), then the result will be printed to
standard error.  It has quirks, though.  For example using print() will
print the line and also the result of print(), which is 'null'.