Fix bug where token stream was pointing in the wrong place after parsing inside square brackets.
Apparently this is not the '99%' scenario, because none of my scripts in
demos/ caught it.  But that's because they express things in ways that
use assemble_expr8()--which didn't have this bug--instead of
assemble_ident_helper().  These two functions are egregious DRY violations
with each other, but short of rewriting the parser entirely, I don't know
how not to require them both.