Reinstat in-place assignment operators for primary elements.
By adding INSTR_LOADATTR which, unlike INSTR_GETATTR, does not remove
the parent object and key from the stack, we can both load an attribute
for use in an operand and then set it again in its parent object.
So now things like "a.b.c++" are possible again.

In the process I also got rid of all those macros in assemble_ident_helper;
now there's just a single helper function in the case of assignment.