De-mangle assemble.c and xptr.c handling of struct xptrvar_t.
This breaks serializer.c.  It wouldn't be that hard to fix, but I might
delete the whole file, since it's a pain to maintain it from code rot,
when we are still so far from even thinking seriously about a
serialization scheme.  I have the whole file disabled with a preprocessor
macro right now.