Rewrite reassemble() to be fast.
Gave up on the dream of being able to re-assemble the human-readable
stuff.  That's what the high-level language is for anyway.  Re-assemble
now only uses the tokenizer when parsing certain .rodata (to prevent some
hairy DRY violations, esp. with strings and bytes).  It is no longer
slower than compiling the source code, and in fact it's blinding fast,
at least for code that is still parsing text.