Do not call BUGCHECK_TYPES in int_cmp.
This is not an .opm method, so it isn't managed by the function in
op.c that calls the float handler if one of the two args is float.
Instead make sure b is at least a real number, then have int_cmp
convert a or b as necessary to perform the operation.