Require private-data arg to .foreach callback only if private data is provided.
Since argument passing in EvilCandy is so much stricter than JavaScript,
there would likely be a lot of ArgumentErrors due to forgetting to declare
a third argument in the callback argument to a .foreach method.  So only
pass two arguments to callback if user did not provide private data, and
pass three arguments if they did (in which case they would have no excuse).