Get rid of INSTR_TERNARY.
Only one of the second or third expressions in a ternary operator will be
evaluated in runtime, so there's no need for INSTR_TERNARY anymore.  This
comes at the small price of a branch instruction and an inability to
reduce an all-const ternary operator into a single INSTR_LOAD_CONST, but
we weren't doing that optimization anyway.  (User codes' ternary
expressions almost never comprise three load-time consts.)