Replace INSTR_LOGICAL_AND/OR with conditional jumps.
This also gets rid of INSTR_B_IF_DEL and just used INSTR_B_IF, since it
also requires a new flag telling VM to keep the condition Object on the
stack if jumping.

To support this in optimization, this patch add a much more aggressive
step to remove unreachable code, should there be any consts in certain
parts of a 'if' or 'while' condition expression.
* Constant true+&& or false+|| will reduce to nothing, leaving the
  remainder of the conditional expression.
* Constant true+|| will cause the entire conditional expression to
  disappear, leaving only the statement it flows into.
* Constant false+&& will cause the entire conditional expression to
  disappear as well as the statement it flows into