Lots more work on reducing unreachable code and modifying-in-place jump instructions.
I have this currently disabled with a macro.  I keep finding corner cases
that need debugging, and since it does little more than improve locality
of reference and remove a B or B_IF instruction here or there, and only
when source code has a hand-written comment in the condition expression,
it's not worth it to enable until it is fully debugged.